+++ components/NotSupported.js
@@ -6,7 +6,7 @@ const NotSupported = () => {
     <div className="flex items-center justify-center w-full m-auto h-screen">
       <div className="shadow-lg rounded-2xl p-4 backdrop-blur-lg bg-white/40 dark:bg-gray-800 max-w-lg">
         <h1 className="dark:text-gray-100 text-gray-900 text-2xl capitalize font-semibold text-center">
-          <code>This route is not supported. Please use a track URL/URI.</code>
+          <code>This route is not supported Please use a track URL/URI</code>
         </h1>
         <div className="flex items-center justify-center gap-4 w-full mt-8">
           <Link